+++ src/plugger/Element/Elements.java
@@ -4,19 +4,19 @@ import java.util.List;
 
 public class Elements{
 	List<Element> element;
-	List<Short> percentage;
-	List<Integer> compression;
+	List<Byte> percentage;
+	List<Byte> compression;
 	
 	public List<Element> getElement() {return element;}
 	public void setElement(List<Element> element) {this.element = element;}
-	public List<Short> getPercentage() {return percentage;}
-	public void setPercentage(List<Short> percentage) {this.percentage = percentage;}
-	public List<Integer> getCompression() {return compression;}
-	public void setCompression(List<Integer> compression) {this.compression = compression;}
+	public List<Byte> getPercentage() {return percentage;}
+	public void setPercentage(List<Byte> percentage) {this.percentage = percentage;}
+	public List<Byte> getCompression() {return compression;}
+	public void setCompression(List<Byte> compression) {this.compression = compression;}
 
 	
 	
-	public Elements(List<Element> element,List<Short> percentage,List<Integer> compression){
+	public Elements(List<Element> element,List<Byte> percentage,List<Byte> compression){
 		this.element=element;
 		this.percentage=percentage;
 		this.compression=compression;
